// frontend/src/api/generated/model/syncRunResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface SyncRunResponse {
  id?: number;
  startedAt?: string;
  finishedAt?: string;
  status?: string;
  accountsTotal?: number;
  accountsOk?: number;
  accountsFailed?: number;
  errorSummary?: string;
}